+++ trans_qt_encode01/trans_qt_encode/encode/prob_est.c
@@ -51,8 +51,7 @@
 
 #define uint unsigned int
 #define uchar unsigned char
-
-void prob_est(Uint32_Dat* n, float *p)
+void prob_est(Uint32_Dat* n, double *p)
 {
 	int nt = 0;
 	int a = 0;
@@ -63,7 +62,7 @@ void prob_est(Uint32_Dat* n, float *p)
 	/* if length(n)<=3 */
 	if (n->len <= 3) {
 		/* p=1-n(2)/n(1);  */
-		*p = 1 - (float)n->dat[1] / n->dat[0];
+		*p = 1 - (double)n->dat[1] / n->dat[0];
 	}
 	else {
 		/*! % 2*nt-term estimation (power to the (1/nt) ): */
@@ -90,11 +89,58 @@ void prob_est(Uint32_Dat* n, float *p)
 			a = c;
 		}
 		/* dx=(a-b)/a; */
-		dx = (float)(a - b) / a;
+		dx = (double)(a - b) / a;
 		/* p= dx/nt  +(1/(2*nt))*(1-1/nt)*(dx*dx);  */
-		*p = dx / nt + (1 / (float)(2 * nt))*(1 - 1 / (float)nt)*(dx*dx);
+		*p = dx / nt + (1 / (double)(2 * nt))*(1 - 1 / (double)nt)*(dx*dx);
 
 	}
 
 	return;
 }
+
+//void prob_est(Uint32_Dat* n, float *p)
+//{
+//	int nt = 0;
+//	int a = 0;
+//	int b = 0;
+//	int c = 0;
+//	float dx = 0;
+//
+//	/* if length(n)<=3 */
+//	if (n->len <= 3) {
+//		/* p=1-n(2)/n(1);  */
+//		*p = 1 - (float)n->dat[1] / n->dat[0];
+//	}
+//	else {
+//		/*! % 2*nt-term estimation (power to the (1/nt) ): */
+//		if (n->len >= 16) {
+//			nt = 4;
+//		}
+//		else {
+//			nt = 2;
+//		}
+//		/* a=sum(n(1:nt)); */
+//		a = 0;
+//		for (int i = 0; i < nt; i++) {
+//			a += n->dat[i];
+//		}
+//		/* b=sum( n(nt+1 : 2*nt) ); */
+//		b = 0;
+//		for (int i = nt; i < 2 * nt; i++) {
+//			b += n->dat[i];
+//		}
+//		/* if b>a,   c=b; b=a; a=c;   end */
+//		if (b > a) {
+//			c = b;
+//			b = a;
+//			a = c;
+//		}
+//		/* dx=(a-b)/a; */
+//		dx = (float)(a - b) / a;
+//		/* p= dx/nt  +(1/(2*nt))*(1-1/nt)*(dx*dx);  */
+//		*p = dx / nt + (1 / (float)(2 * nt))*(1 - 1 / (float)nt)*(dx*dx);
+//
+//	}
+//
+//	return;
+//}
